add unreachable item summary to dominator output
move dominators analysis into submodule

split dominators emitting code into submodule

formatting fixes

remove extraneous module in emit

add comment explaining scoped borrow
