--- analyze/analyses/dominators.rs
@@ -1,257 +1,0 @@
-use std::collections::BTreeMap;
-use std::io;
-
-use csv;
-use regex;
-
-use formats::json;
-use formats::table::{Align, Table};
-use twiggy_ir as ir;
-use twiggy_opt as opt;
-use twiggy_traits as traits;
-
-struct DominatorTree {
-    tree: BTreeMap<ir::Id, Vec<ir::Id>>,
-    items: Vec<ir::Id>,
-    opts: opt::Dominators,
-}
-
-impl traits::Emit for DominatorTree {
-    #[cfg(feature = "emit_text")]
-    fn emit_text(&self, items: &ir::Items, dest: &mut io::Write) -> Result<(), traits::Error> {
-        let mut table = Table::with_header(vec![
-            (Align::Right, "Retained Bytes".to_string()),
-            (Align::Right, "Retained %".to_string()),
-            (Align::Left, "Dominator Tree".to_string()),
-        ]);
-
-        let opts = &self.opts;
-        let mut row = 0 as u32;
-
-        fn recursive_add_rows(
-            table: &mut Table,
-            items: &ir::Items,
-            dominator_tree: &BTreeMap<ir::Id, Vec<ir::Id>>,
-            depth: u32,
-            mut row: &mut u32,
-            opts: &opt::Dominators,
-            id: ir::Id,
-        ) {
-            assert_eq!(id == items.meta_root(), depth == 0);
-
-            if *row == opts.max_rows() {
-                return;
-            }
-
-            if depth > opts.max_depth() {
-                return;
-            }
-
-            if depth > 0 {
-                let item = &items[id];
-
-                let size = items.retained_size(id);
-                let size_percent = (f64::from(size)) / (f64::from(items.size())) * 100.0;
-
-                let mut label =
-                    String::with_capacity(depth as usize * 4 + item.name().len() + "⤷ ".len());
-                for _ in 2..depth {
-                    label.push_str("    ");
-                }
-                if depth != 1 {
-                    label.push_str("  ⤷ ");
-                }
-                label.push_str(item.name());
-
-                table.add_row(vec![
-                    size.to_string(),
-                    format!("{:.2}%", size_percent),
-                    label,
-                ]);
-            }
-
-            if let Some(children) = dominator_tree.get(&id) {
-                let mut children = children.to_vec();
-                children.sort_by(|a, b| items.retained_size(*b).cmp(&items.retained_size(*a)));
-                for child in children {
-                    *row += 1;
-                    recursive_add_rows(
-                        table,
-                        items,
-                        dominator_tree,
-                        depth + 1,
-                        &mut row,
-                        &opts,
-                        child,
-                    );
-                }
-            }
-        }
-
-        for id in &self.items {
-            let start_depth = if *id == items.meta_root() { 0 } else { 1 };
-            recursive_add_rows(
-                &mut table,
-                items,
-                &self.tree,
-                start_depth,
-                &mut row,
-                &opts,
-                *id,
-            );
-        }
-
-        write!(dest, "{}", &table)?;
-        Ok(())
-    }
-
-    #[cfg(feature = "emit_json")]
-    fn emit_json(&self, items: &ir::Items, dest: &mut io::Write) -> Result<(), traits::Error> {
-        fn recursive_add_children(
-            items: &ir::Items,
-            opts: &opt::Dominators,
-            dominator_tree: &BTreeMap<ir::Id, Vec<ir::Id>>,
-            id: ir::Id,
-            obj: &mut json::Object,
-        ) -> Result<(), traits::Error> {
-            let item = &items[id];
-
-            obj.field("name", item.name())?;
-
-            let size = item.size();
-            let size_percent = f64::from(size) / f64::from(items.size()) * 100.0;
-            obj.field("shallow_size", size)?;
-            obj.field("shallow_size_percent", size_percent)?;
-
-            let size = items.retained_size(id);
-            let size_percent = f64::from(size) / f64::from(items.size()) * 100.0;
-            obj.field("retained_size", size)?;
-            obj.field("retained_size_percent", size_percent)?;
-
-            // TODO: this needs to do the filtering like how text formatting
-            // does, but it would be nice to push that earlier, like `top` does.
-
-            if let Some(children) = dominator_tree.get(&id) {
-                let mut children = children.to_vec();
-                children.sort_by(|a, b| items.retained_size(*b).cmp(&items.retained_size(*a)));
-
-                let mut arr = obj.array("children")?;
-                for child in children {
-                    let mut obj = arr.object()?;
-                    recursive_add_children(items, opts, dominator_tree, child, &mut obj)?;
-                }
-            }
-
-            Ok(())
-        }
-
-        let mut obj = json::object(dest)?;
-        for curr_id in &self.items {
-            recursive_add_children(items, &self.opts, &self.tree, *curr_id, &mut obj)?;
-        }
-
-        Ok(())
-    }
-
-    #[cfg(feature = "emit_csv")]
-    fn emit_csv(&self, items: &ir::Items, dest: &mut io::Write) -> Result<(), traits::Error> {
-        let mut wtr = csv::Writer::from_writer(dest);
-        fn recursive_add_children(
-            items: &ir::Items,
-            opts: &opt::Dominators,
-            dominator_tree: &BTreeMap<ir::Id, Vec<ir::Id>>,
-            id: ir::Id,
-            wtr: &mut csv::Writer<&mut io::Write>,
-        ) -> Result<(), traits::Error> {
-            #[derive(Serialize, Debug)]
-            #[serde(rename_all = "PascalCase")]
-            struct CsvRecord {
-                id: u64,
-                name: String,
-                shallow_size: u32,
-                shallow_size_percent: f64,
-                retained_size: u32,
-                retained_size_percent: f64,
-                immediate_dominator: u64,
-            }
-
-            let item = &items[id];
-            let (size, size_percent) = (
-                item.size(),
-                f64::from(item.size()) / f64::from(items.size()) * 100.0,
-            );
-            let (retained_size, retained_size_percent) = (
-                items.retained_size(id),
-                f64::from(items.retained_size(id)) / f64::from(items.size()) * 100.0,
-            );
-            let idom = if let Some(idom) = items.immediate_dominators().get(&id) {
-                idom.serializable()
-            } else {
-                id.serializable()
-            };
-
-            let rc = CsvRecord {
-                id: item.id().serializable(),
-                name: item.name().to_string(),
-                shallow_size: size,
-                shallow_size_percent: size_percent,
-                retained_size,
-                retained_size_percent,
-                immediate_dominator: idom,
-            };
-
-            wtr.serialize(rc)?;
-            wtr.flush()?;
-            if let Some(children) = dominator_tree.get(&id) {
-                let mut children = children.to_vec();
-                children.sort_by(|a, b| items.retained_size(*b).cmp(&items.retained_size(*a)));
-                for child in children {
-                    recursive_add_children(items, opts, dominator_tree, child, wtr)?;
-                }
-            }
-            Ok(())
-        }
-
-        recursive_add_children(items, &self.opts, &self.tree, items.meta_root(), &mut wtr)?;
-        Ok(())
-    }
-}
-
-/// Compute the dominator tree for the given IR graph.
-pub fn dominators(
-    items: &mut ir::Items,
-    opts: &opt::Dominators,
-) -> Result<Box<traits::Emit>, traits::Error> {
-    items.compute_dominator_tree();
-    items.compute_dominators();
-    items.compute_retained_sizes();
-    items.compute_predecessors();
-
-    let arguments = opts.items();
-    let dominator_items = if arguments.is_empty() {
-        vec![items.meta_root()]
-    } else if opts.using_regexps() {
-        let regexps = regex::RegexSet::new(arguments)?;
-        let mut sorted_items: Vec<_> = items
-            .iter()
-            .filter(|item| regexps.is_match(&item.name()))
-            .map(|item| item.id())
-            .collect();
-        sorted_items.sort_by_key(|id| -i64::from(items.retained_size(*id)));
-        sorted_items
-    } else {
-        arguments
-            .iter()
-            .filter_map(|name| items.get_item_by_name(name))
-            .map(|item| item.id())
-            .collect()
-    };
-
-    let tree = DominatorTree {
-        tree: items.dominator_tree().clone(),
-        items: dominator_items,
-        opts: opts.clone(),
-    };
-
-    Ok(Box::new(tree) as Box<traits::Emit>)
-}
