+++ analyze/analyses/mod.rs
@@ -4,3 +4,5 @@ pub mod garbage;
 pub mod monos;
 pub mod paths;
 pub mod top;
+
+mod utils;
