+++ twiggy/tests/all/expectations/dominators_csv_does_not_summarize_garbage_if_all_items_are_reachable
@@ -1,0 +1,24 @@
+Id,Name,ShallowSize,ShallowSizePercent,RetainedSize,RetainedSizePercent,ImmediateDominator
+18446744073709551615,<meta root>,0,0,144,100,18446744073709551615
+17179869184,"""function names"" subsection",46,31.944444444444443,46,31.944444444444443,18446744073709551615
+8589934594,"export ""woof""",7,4.861111111111112,22,15.277777777777779,18446744073709551615
+4294967299,func[3],1,0.6944444444444444,15,10.416666666666668,8589934594
+12884901891,woof,8,5.555555555555555,14,9.722222222222223,4294967299
+4294967296,func[0],1,0.6944444444444444,6,4.166666666666666,12884901891
+12884901888,calledOnce,5,3.4722222222222223,5,3.4722222222222223,4294967296
+8589934592,"export ""awoo""",7,4.861111111111112,13,9.027777777777777,18446744073709551615
+4294967300,func[4],1,0.6944444444444444,6,4.166666666666666,8589934592
+12884901892,awoo,5,3.4722222222222223,5,3.4722222222222223,4294967300
+17179869185,"""local names"" subsection",13,9.027777777777777,13,9.027777777777777,18446744073709551615
+25769803775,wasm magic bytes,8,5.555555555555555,8,5.555555555555555,18446744073709551615
+8589934593,"export ""bark""",7,4.861111111111112,7,4.861111111111112,18446744073709551615
+21474836479,custom section 'name' headers,7,4.861111111111112,7,4.861111111111112,18446744073709551615
+4294967297,func[1],1,0.6944444444444444,6,4.166666666666666,18446744073709551615
+12884901889,calledTwice,5,3.4722222222222223,5,3.4722222222222223,4294967297
+4294967298,func[2],1,0.6944444444444444,6,4.166666666666666,18446744073709551615
+12884901890,bark,5,3.4722222222222223,5,3.4722222222222223,4294967298
+0,type[0]: () -> i32,4,2.7777777777777777,4,2.7777777777777777,18446744073709551615
+4294967295,type section headers,3,2.083333333333333,3,2.083333333333333,18446744073709551615
+8589934591,function section headers,3,2.083333333333333,3,2.083333333333333,18446744073709551615
+12884901887,export section headers,3,2.083333333333333,3,2.083333333333333,18446744073709551615
+17179869183,code section headers,3,2.083333333333333,3,2.083333333333333,18446744073709551615
