+++ twiggy/tests/all/expectations/dominators_json_prints_multiple_root_items
@@ -1,0 +1,1 @@
+{"items":[{"name":"calledOnce","shallow_size":5,"shallow_size_percent":3.4722222222222223,"retained_size":5,"retained_size_percent":3.4722222222222223},{"name":"calledTwice","shallow_size":5,"shallow_size_percent":3.4722222222222223,"retained_size":5,"retained_size_percent":3.4722222222222223}]}
