+++ twiggy/tests/all/expectations/dominators_wee_alloc_csv
@@ -46,3 +46,4 @@ Id,Name,ShallowSize,ShallowSizePercent,RetainedSize,RetainedSizePercent,Immediat
 17179869183,memory section headers,3,0.10649627263045794,3,0.10649627263045794,18446744073709551615
 21474836479,export section headers,3,0.10649627263045794,3,0.10649627263045794,18446744073709551615
 25769803775,element section headers,3,0.10649627263045794,3,0.10649627263045794,18446744073709551615
+,[3 Unreachable Items],1084,38.480653177138805,1084,38.480653177138805,
