+++ twiggy/tests/all/expectations/dominators_wee_alloc_subtree_json
@@ -1,1 +1,1 @@
-{"name":"hello","shallow_size":164,"shallow_size_percent":5.8217962371317,"retained_size":164,"retained_size_percent":5.8217962371317}
+{"items":[{"name":"hello","shallow_size":164,"shallow_size_percent":5.8217962371317,"retained_size":164,"retained_size_percent":5.8217962371317}]}
